hw8 fix all

+++ src/com/urise/webapp/storage/SortedArrayStorage.java
@@ -3,7 +3,6 @@ package com.urise.webapp.storage;
 import com.urise.webapp.model.Resume;
 
 import java.util.Arrays;
-import java.util.List;
 
 public class SortedArrayStorage extends AbstractArrayStorage {
 
@@ -15,8 +14,8 @@ public class SortedArrayStorage extends AbstractArrayStorage {
 
     @Override
     public void addToArray(Resume r) {
-        int index = (int) searchKey(r.getUuid());
-        int insertIndex = (-index) - 1;
+        int index = searchKey(r.getUuid());
+        int insertIndex = -index - 1;
         if (insertIndex < storage.length - 1) {
             System.arraycopy(storage, insertIndex, storage, insertIndex + 1, size - insertIndex);
         }
